requirements.txt
numpy
pandas
scikit-learn>=1.3
torch>=2.0
matplotlib
pyyaml
shap
networkx